// src/api/models/user_meta.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { user_images } from './user_images';
export type user_meta = {
  /**
   * MyAnimeList Username
   */
  username?: string;
  /**
   * MyAnimeList Profile URL
   */
  url?: string;
  images?: user_images;
};